Added Buttons
Added three classes: Button, ButtonAction, and
ButtonAction_IterateNumber. The first two define a simple and generic
button that will run the ButtonAction's action when the button is
pressed. The last extends the ButtonAction and defines the action to add
a given interval to a TextMesh's current value. This is then used on the
start menu to select how many trials to run.