Added Trial
Made a new scene for trials and made a class Trial which is given a
TextMesh to indicate the target, target sprite, target color, list of
possible colors, how many objects to show during the trial, and time
intervals for how long to show and object and how long between objects.
It sets up the trial by assigning target different colors from the color
pool and placing the target color in one of them with 1/(number of
objects to show) chance of not appearing at all.