Added input to Trial
Added the ability to read in the space input and keep track of the
number of times it was pressed, the object it was last pressed on, and
the time spent on that object before space was pressed.

Also updated comments and added the base Trial Manager for next update.